refactor(tests): fold test_query_dispatch/runner_base into test_select_runner; split test_aggregate.cpp
Consolidates three test helper headers (test_query_dispatch.h,
test_query_runner_base.h, test_select_runner.h) into one, and splits
the 1742-line test_aggregate.cpp into five focused files each under
600 lines with no duplicate blocks.

New files:
- test_select_runner.h — absorbs dispatch_field, QueryRunnerBase, SelectQueryRunnerBase
- tests/query/test_aggregate_fixture.h — shared AggregateTest fixture + insert helpers
- tests/query/test_aggregate_groupby.cpp — GROUP BY + COUNT(DISTINCT) tests
- tests/query/test_aggregate_optional.cpp — NULL/Optional + Negative number tests
- tests/query/test_aggregate_combined.cpp — FullChain + GROUP BY modifiers tests
- tests/query/test_aggregate_having.cpp — HAVING + GroupByOrderByTest + HAVING+modifiers tests
- tests/query/test_where_complex.cpp — ComplexWhereTest moved from test_where.cpp

Deleted: test_query_dispatch.h, test_query_runner_base.h

All 1901 tests pass under ninja-asan-ubsan.
